Allow users to supply Sessions or a Session class
The previous workflow introduced unneeded overhead instantiating an
extra Session for each Poloniex object over PoloniexPublic, and
making a new session for each object.

Not only that, but the default requests Session is not threadsafe.

In order to be compliant to be threadsafe, one would have to use
a custom Session class/instance, and reuse that instance per thread
which may have their own Poloniex objects.